requirements.txt
Flask==2.3.2
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
requests==2.31.0
tensorflow==2.12.0
